+++ package.json
@@ -6,12 +6,11 @@
   "scripts": {
     "develop": "vite",
     "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "lint": "eslint src --max-warnings 0",
     "preview": "vite preview",
-    "jest": "jest",
+    "test": "vitest",
     "prettier": "prettier --ignore-path ./.prettierignore src --check",
     "prettier:fix": "prettier --ignore-path ./.prettierignore src --write",
-    "test": "jest",
     "check-dependencies": "npm-check",
     "cypress": "cypress run --component --headed",
     "cypress:open": "cypress open",
@@ -25,7 +24,6 @@
     "@ukic/web-components": "^3.13.0",
     "cypress-real-events": "^1.14.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest-environment-jsdom": "^29.7.0",
     "prettier": "^3.3.3",
     "react-router-dom": "^6.26.0",
     "ts-node": "^10.9.2",
@@ -35,27 +33,26 @@
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.4",
     "@babel/preset-typescript": "^7.24.7",
+    "@eslint/js": "^9.39.2",
     "@testing-library/dom": "^10.4.0",
-    "@testing-library/jest-dom": "^6.4.8",
+    "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
-    "@types/jest": "^29.5.12",
+    "@types/node": "^25.0.9",
     "@types/react": "^19.1.6",
     "@types/react-dom": "^19.1.5",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "babel-jest": "^29.7.0",
+    "@vitejs/plugin-react": "^4.7.0",
     "cypress": "^14.5.0",
-    "eslint": "^8.45.0",
+    "eslint": "^9.18.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.3",
-    "jest": "^29.7.0",
+    "globals": "^17.0.0",
+    "jsdom": "^27.4.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "shadow-dom-testing-library": "^1.11.2",
-    "ts-jest": "^29.2.4",
     "typescript": "^5.5.4",
+    "typescript-eslint": "^8.53.0",
     "vite": "^5.4.2",
     "vitest": "^2.1.8"
   },
