+++ src/components/Home/__tests__/Home.test.tsx
@@ -10,7 +10,7 @@ describe("Home component", () => {
   });
 
   afterEach(() => {
-    jest.restoreAllMocks();
+    vi.restoreAllMocks();
   });
   it("renders", async () => {
     expect(container).not.toBeNull();
