+++ src/components/Info/__tests__/Info.test.tsx
@@ -10,7 +10,7 @@ describe("Info component", () => {
   });
 
   afterEach(() => {
-    jest.restoreAllMocks();
+    vi.restoreAllMocks();
   });
   it("renders", async () => {
     expect(container).not.toBeNull();
