+++ src/setupTests.ts
@@ -1,0 +1,5 @@
+import "@testing-library/jest-dom";
+import { defineCustomElements } from "@ukic/web-components/loader";
+
+// Register UKIC web components
+defineCustomElements();
